dom/media/tests/mochitest/test_setIdentityProvider.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8"/>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="head.js"></script>
  <script type="application/javascript" src="pc.js"></script>
  <script type="application/javascript" src="templates.js"></script>
</head>
<body>
<pre id="test">
<script type="application/javascript">
  createHTML({
    title: "setIdentityProvider leads to peerIdentity and assertions in SDP"
  });

var test;
function theTest() {
  test = new PeerConnectionTest();
  test.setMediaConstraints([{audio: true}], [{audio: true}]);
  test.setIdentityProvider(test.pcLocal, 'test1.example.com', 'idp.html', 'nobody');
  test.setIdentityProvider(test.pcRemote, 'test2.example.com', 'idp.html', 'nobody');
  test.chain.append([
  [
    "PEER_IDENTITY_IS_SET_CORRECTLY",
    function(test) {
      var outstanding = 0;
      // we have to wait for the identity result in order to get the actual
      // identity information, since the call will complete before the identity
      // provider has a chance to finish verifying... that's OK, but it makes
      // testing more difficult
      function checkOrSetupCheck(pc, pfx, idp, name) {
        if (pc.peerIdentity) {
          is(pc.peerIdentity.idp, idp, pfx + "IdP is correct");
          is(pc.peerIdentity.name, name + "@" + idp, pfx + "identity is correct");
        } else {
          ++outstanding;
          pc.onidentityresult = function checkIdentity(e) {
            console.log("Got result", e, e.assertion, e.peerIdentity);
            if (e.assertion) {
              return;
            }
            ok(e.peerIdentity, pfx + "identity result contains identity");
            is(e.peerIdentity, e.target.peerIdentity, pfx + "identity result is on PC");
            is(e.peerIdentity.idp, idp, pfx + "IdP is correct");
            is(e.peerIdentity.name, name + "@" + idp, pfx + "identity is correct");
            --outstanding;
            if (outstanding <= 0) {
              test.next();
            }
          };
        }
      }

      checkOrSetupCheck(test.pcLocal._pc, "local: ", "test2.example.com", "someone");
      checkOrSetupCheck(test.pcRemote._pc, "remote: ", "test1.example.com", "someone");
      if (outstanding <= 0) {
        test.next();
      }
    }
  ],
  [
    "OFFERS_AND_ANSWERS_INCLUDE_IDENTITY",
    function(test) {
      ok(test.pcLocal._last_offer.sdp.contains("a=identity"), "a=identity is in the offer SDP");
      ok(test.pcRemote._last_answer.sdp.contains("a=identity"), "a=identity is in the answer SDP");
      test.next();
    }
  ],
  [
    "DESCRIPTIONS_CONTAIN_IDENTITY",
    function(test) {
      ok(test.pcLocal.localDescription.sdp.contains("a=identity"),
                         "a=identity is in the local copy of the offer");
      ok(test.pcRemote.localDescription.sdp.contains("a=identity"),
                         "a=identity is in the remote copy of the offer");
      ok(test.pcLocal.remoteDescription.sdp.contains("a=identity"),
                         "a=identity is in the local copy of the answer");
      ok(test.pcRemote.remoteDescription.sdp.contains("a=identity"),
                         "a=identity is in the remote copy of the answer");
      test.next();
    }
  ]
  ]);
  test.run();
}
runTest(theTest);



</script>
</pre>
</body>
</html>
